backend/pyproject.toml
[project]
name = "shortcuts-generator-backend"
version = "0.1.0"
description = "FastAPI backend for generating and signing Apple Shortcuts."
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
    "google-genai>=1.0.0",
    "httpx>=0.27.0",
    "boto3>=1.35.0",
    "python-dotenv>=1.0.0",
    "PyJWT[crypto]>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py39"
